+++ programs/mango-v4/tests/program_test/mango_client.rs
@@ -5174,7 +5174,7 @@ impl<T: ClientInstruction> ClientInstruction for HealthAccountSkipping<T> {
 pub struct SequenceCheckInstruction {
     pub account: Pubkey,
     pub owner: TestKeypair,
-    pub expected_sequence_number: u64,
+    pub expected_sequence_number: u8,
 }
 #[async_trait::async_trait(?Send)]
 impl ClientInstruction for SequenceCheckInstruction {
